NET-276 NNTPClient has problems with group listings for large groups.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/trunk@1079634 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/examples/nntp/NNTPUtils.java
@@ -41,7 +41,7 @@ public class NNTPUtils {
      * @return Article[] An array of Article
      * @throws IOException
      */
-    public  static List<Article> getArticleInfo(NNTPClient client, int lowArticleNumber, int highArticleNumber)
+    public  static List<Article> getArticleInfo(NNTPClient client, long lowArticleNumber, long highArticleNumber)
     throws IOException {
         Reader reader = null;
         List<Article> articles = new ArrayList<Article>();
